added backend for package booking
users will get redirected to the agents page when clicking the book button on the package details page. Selecting an agent will create a connection between the agent, user, and package. Meanwhile, agents will get a dropdown to select which user they want to book the package for. This will only work if there is a previous connection initiated by a user. The package details page is still not ready. The book package page was deleted, its functionality merged with the package details page.